+++ js/consts.js
@@ -1,0 +1,47 @@
+const MAX_COMMENTS = 2;
+const AVATAR_COUNT = 6;
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const MAX_COMMENT_ID = 300;
+const PHOTOS_COUNT = 25;
+const DESCRIPTIONS = [
+  'Фотография на фоне гор',
+  'Фотография на фоне моря',
+  'Фотография в городе',
+  'Селфи на фоне деревьев',
+  'Кот сидит на балконе',
+  'Группа людей в аудитории',
+  'Фотография еды в кафе',
+  'Неопознанный летающий объект',
+  'Закат солнца на фоне реки',
+  'Быт в деревне'
+];
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+const FIRST_NAMES = [
+  'Олег',
+  'Артём',
+  'Вася',
+  'Пётр',
+  'Саша',
+  'Алёна',
+  'Ольга',
+  'Софья',
+  'Наталья'
+];
+
+export {MAX_COMMENTS};
+export {AVATAR_COUNT};
+export {MIN_LIKES};
+export {MAX_LIKES};
+export {MAX_COMMENT_ID};
+export {PHOTOS_COUNT};
+export {DESCRIPTIONS};
+export {MESSAGES};
+export {FIRST_NAMES};
