--- js/random-number.js
@@ -1,5 +1,0 @@
-function getRandomNumber(maxValue) {
-  return Math.floor(Math.random() * (maxValue + 1));
-}
-
-export {getRandomNumber};
